shopee/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
awsebcli==3.20.10
beautifulsoup4==4.12.3
blessed==1.20.0
botocore==1.31.85
cache==1.0.3
cement==2.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.3
Django==4.2.6
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
hiredis==2.3.2
idna==3.6
jmespath==1.0.1
pathspec==0.10.1
Pillow==10.1.0
psycopg==3.1.12
psycopg2==2.9.9
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
termcolor==1.1.0
typing-extensions==4.8.0
urllib3==1.26.18
wcwidth==0.1.9
